refactor: isolate studio panel sections for better performance
- Replace monolithic StudioPanel (604 lines) with context-based architecture
- Create StudioContext for shared state (projectId, signals, validSourceIds)
- Extract 18 self-contained feature sections (Audio, Quiz, Video, etc.)
- Each section owns its own state - only re-renders when its data changes
- Eliminate prop drilling (126 props -> 0)
- Use memoized Set for O(1) source filtering instead of O(n²) nested loops
- Remove unused components: StudioGeneratedContent, StudioModals, StudioProgressIndicators

Additional fixes:
- Fix audio download opening new tab instead of downloading (cross-origin blob fetch)
- Fix mind map zoom icons (use MagnifyingGlassPlus/Minus)
- Fix mind map zoom reset on node expand (only fitView on initial load)

Performance impact:
- Before: Any state change re-rendered all 18 features
- After: Only affected section re-renders
