+++ frontend/src/components/studio/sections/InfographicSection.tsx
@@ -1,0 +1,64 @@
+/**
+ * InfographicSection Component
+ * Educational Note: Self-contained section for infographic generation.
+ */
+
+import React, { useEffect, useCallback } from 'react';
+import { useStudioContext, useFilteredJobs } from '../StudioContext';
+import { useInfographicGeneration } from '../infographic/useInfographicGeneration';
+import { InfographicListItem } from '../infographic/InfographicListItem';
+import { InfographicProgressIndicator } from '../infographic/InfographicProgressIndicator';
+import { InfographicViewerModal } from '../infographic/InfographicViewerModal';
+
+export const InfographicSection: React.FC = () => {
+  const { projectId, registerGenerationHandler } = useStudioContext();
+
+  const {
+    savedInfographicJobs,
+    currentInfographicJob,
+    isGeneratingInfographic,
+    viewingInfographicJob,
+    setViewingInfographicJob,
+    loadSavedJobs,
+    handleInfographicGeneration,
+  } = useInfographicGeneration(projectId);
+
+  const filteredJobs = useFilteredJobs(savedInfographicJobs);
+
+  useEffect(() => {
+    loadSavedJobs();
+  }, [projectId]); // eslint-disable-line react-hooks/exhaustive-deps
+
+  const handleGenerate = useCallback(async (signal: Parameters<typeof handleInfographicGeneration>[0]) => {
+    await handleInfographicGeneration(signal);
+  }, [handleInfographicGeneration]);
+
+  useEffect(() => {
+    registerGenerationHandler('infographics', handleGenerate);
+  }, [registerGenerationHandler, handleGenerate]);
+
+  if (filteredJobs.length === 0 && !isGeneratingInfographic) {
+    return null;
+  }
+
+  return (
+    <>
+      {isGeneratingInfographic && (
+        <InfographicProgressIndicator currentInfographicJob={currentInfographicJob} />
+      )}
+
+      {filteredJobs.map((job) => (
+        <InfographicListItem
+          key={job.id}
+          job={job}
+          onClick={() => setViewingInfographicJob(job)}
+        />
+      ))}
+
+      <InfographicViewerModal
+        viewingInfographicJob={viewingInfographicJob}
+        onClose={() => setViewingInfographicJob(null)}
+      />
+    </>
+  );
+};
