+++ frontend/src/components/studio/sections/index.ts
@@ -1,0 +1,24 @@
+/**
+ * Studio Sections Index
+ * Educational Note: Each section is self-contained and manages its own state.
+ * This isolation prevents unnecessary re-renders across the Studio panel.
+ */
+
+export { AudioSection } from './AudioSection';
+export { QuizSection } from './QuizSection';
+export { FlashCardSection } from './FlashCardSection';
+export { MindMapSection } from './MindMapSection';
+export { AdSection } from './AdSection';
+export { SocialSection } from './SocialSection';
+export { InfographicSection } from './InfographicSection';
+export { EmailSection } from './EmailSection';
+export { WebsiteSection } from './WebsiteSection';
+export { ComponentSection } from './ComponentSection';
+export { VideoSection } from './VideoSection';
+export { FlowDiagramSection } from './FlowDiagramSection';
+export { WireframeSection } from './WireframeSection';
+export { PresentationSection } from './PresentationSection';
+export { PRDSection } from './PRDSection';
+export { MarketingStrategySection } from './MarketingStrategySection';
+export { BlogSection } from './BlogSection';
+export { BusinessReportSection } from './BusinessReportSection';
